+++ backend/nextcanope/src/main/java/com/veolia/nextcanope/service/LayerService.java
@@ -32,7 +32,7 @@ public class LayerService {
      * @param tileNumber The tile number to search for in the database.
      * @return The equipment tile as a string, associated with the given key and tile number.
      */
-    public String getLayerTile(String key, Long tileNumber, Integer userId) {
+    public String getLayerTile(String key, Long tileNumber, Long userId) {
         return layerRepositoryImpl.getLayerTile(key, tileNumber, userId);
     }
 }
